+++ packages/store/src/public_api.ts
@@ -4,7 +4,7 @@ export { Store } from './store';
 export { State } from './decorators/state';
 export { Select } from './decorators/select/select';
 export { SelectorOptions } from './decorators/selector-options';
-export { Actions } from './actions-stream';
+export { Actions, ActionContext, ActionStatus } from './actions-stream';
 
 // TODO: v4 - We need to come up with an alternative api to exposing this metadata
 //   because it is currently used by the following (after a github search):